Move tranding location to a different file
