+++ frontend/src/components/home/trending.css
@@ -1,7 +1,7 @@
 .home-page-container{
-  display: flex;
+  /* display: flex;
   flex-direction: column;
-  justify-content: center;
+  justify-content: center; */
   align-items: center;
 
   margin: 0 auto;
